implement fs event api for kit
