+++ packages/kit/index.ts
@@ -1,5 +1,3 @@
 export * from '@volar/language-service';
 export * from './lib/createFormatter';
-export * from './lib/createLinter';
-export * from './lib/createProject';
-export * from './lib/createServiceEnvironment';
+export * from './lib/createChecker';
