+++ packages/kit/lib/createFormatter.ts
@@ -1,19 +1,19 @@
 import { FormattingOptions, Language, Service, createFileProvider, createLanguageService } from '@volar/language-service';
 import * as ts from 'typescript';
 import { TextDocument } from 'vscode-languageserver-textdocument';
-import { createServiceEnvironment } from '..';
+import { createServiceEnvironment } from './createServiceEnvironment';
 
 export function createFormatter(
 	languages: Language[],
-	services: Service[],
-	settings: any = {},
+	services: Service[]
 ) {
 
 	let fakeUri = 'file:///dummy.txt';
 	let fakeFileName = '/dummy.txt';
+	let settings = {};
 
+	const env = createServiceEnvironment(() => settings);
 	const fileProvider = createFileProvider(languages, () => { });
-	const env = createServiceEnvironment(settings);
 	const service = createLanguageService(
 		{ typescript: ts as any },
 		services,
@@ -21,9 +21,20 @@ export function createFormatter(
 		{ fileProvider },
 	);
 
-	return { formatCode };
+	return {
+		// apis
+		format,
 
-	async function formatCode(content: string, languageId: string, options: FormattingOptions): Promise<string> {
+		// settings
+		get settings() {
+			return settings;
+		},
+		set settings(v) {
+			settings = v;
+		},
+	};
+
+	async function format(content: string, languageId: string, options: FormattingOptions): Promise<string> {
 
 		fileProvider.updateSource(fakeFileName, ts.ScriptSnapshot.fromString(content), languageId);
